+++ app/src/main/java/de/baumann/browser/activity/Settings_Menu.java
@@ -82,7 +82,6 @@ public class Settings_Menu extends AppCompatActivity {
         itemTouchHelper.attachToRecyclerView(recyclerView);
     }
 
-
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.menu_help, menu);